parser: Merge skipItem and leaveNestingLevel.
Breaking change: skipItem, when exiting an Object/Array and reaching
depth zero, will now return the token `ObjectEnd`/ `ArrayEnd` and not
`peekSignificant()`. The End-Token is still consumed.

This is because when using leaveNestingLevel with starting depth>1, it
is pretty useful to find out the last type of compound (Object vs Array)
you left as a sanity check.