fromJSON: Do not enforce order of template arguments
Since fromJSON is only called via IFTI, whether the JT or P template
argument comes first or second doesn't matter.

It only throws an error when the function is inspected to check whether
JT has the `ref` storage class. The template has to be instantiated by
that (without IFTI) and thus determine the correct order.

Of course it would be possible to just define a required order and
document that, but for better error messages, it would still be useful
to determine if the programmer made a mistake it. But after the
determination has been done, it's trivial to just write the check order
agnostic too.